Add files via upload
This was my first try for a PCA in Rstudio. You can also find the excel tables for positive and negative correlations. Please note that the both excel tables hold ALL correlations. The PDF shows a picture of positive (blue) and negative (pink) correlations - you can zoom in to read the names on the knots. If it´s not helpful, please delete it - no worries :)